Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 int main(){
-    float a;
-    scanf("%f",&a);
+    double a;
+    scanf("%lf",&a);
     if (a == 0){
-printf("");
+    printf("");
     }
     else
-{printf("You entered: %.4f",a);
+    {printf("You entered: %.4lf",a);
    }   return 0;
 }
